Renamed pwm_rx to rx_pwm
