+++ src/main/drivers/rx_pwm.c
@@ -36,7 +36,7 @@
 #include "pwm_output.h"
 #include "pwm_mapping.h"
 
-#include "pwm_rx.h"
+#include "rx_pwm.h"
 
 #define DEBUG_PPM_ISR
