+++ src/main/fc/config.h
@@ -21,8 +21,8 @@
 #include <stdbool.h>
 #include "common/time.h"
 #include "config/parameter_group.h"
-#include "drivers/pwm_rx.h"
 #include "drivers/adc.h"
+#include "drivers/rx_pwm.h"
 #include "fc/stats.h"
 
 #define MAX_PROFILE_COUNT 3
